Added feature "find_file_at_point"
This feature allows to place the cursor at any location inside
the document and call the function. Now it will try to extract
a meaningful filename and try to locate the file inside the open
projects.

In any case it will present an input view with the full path and ask
the user to open this file, if the propsed file exists it will show
the located path, else only the base path of the file and the extracted
partial filename.

The shortcut for this feature is again borrowed form Emacs

Alt+x f f a p